feat: introduce new specifications and refactor project structure
- add `MultiOrderSpecification<T>` class to handle multiple ordering criteria.
- add `NotSpecification<T>` class to negate another specification.
- move existing `NotSpecification<T>` from `AlwaysTrueSpecification.cs` to its own file.
- remove `MultiOrderSpecification<T>` from `IOrderSpecification.cs` and place it in its own file.
- rename test project from `Candoumbe.DataAccess.Tests` to `Candoumbe.DataAccess.UnitTests`.
- update solution configuration to reflect the new project structure.
- adjust Stryker configuration files to remove concurrency settings.
