+++ Candoumbe.DataAccess.sln
@@ -8,8 +8,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{91B90448-4
 		stryker-config.json = stryker-config.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Candoumbe.DataAccess.UnitTests", "test\Candoumbe.DataAccess.Tests\Candoumbe.DataAccess.UnitTests.csproj", "{C572DA49-1F54-4884-AC07-851CF64264DD}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A960764B-242B-4163-B67B-5B28B8206E50}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Candoumbe.DataAccess.RavenDb", "src\Candoumbe.DataAccess.RavenDb\Candoumbe.DataAccess.RavenDb.csproj", "{3B4BF7C5-73FB-4BF8-B64E-A41BEA003266}"
@@ -60,16 +58,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Candoumbe.DataAccess.Abstra
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Candoumbe.DataAccess.Abstractions.UnitTests", "test\Candoumbe.DataAccess.Abstractions.UnitTests\Candoumbe.DataAccess.Abstractions.UnitTests.csproj", "{BB2CB8B1-AA48-4361-B6FE-558FE40E8E9C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Candoumbe.DataAccess.UnitTests", "test\Candoumbe.DataAccess.UnitTests\Candoumbe.DataAccess.UnitTests.csproj", "{E47B721F-EFCA-4302-AAB1-6AE97A62F813}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C572DA49-1F54-4884-AC07-851CF64264DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C572DA49-1F54-4884-AC07-851CF64264DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C572DA49-1F54-4884-AC07-851CF64264DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C572DA49-1F54-4884-AC07-851CF64264DD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3B4BF7C5-73FB-4BF8-B64E-A41BEA003266}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3B4BF7C5-73FB-4BF8-B64E-A41BEA003266}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B4BF7C5-73FB-4BF8-B64E-A41BEA003266}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -100,12 +96,15 @@ Global
 		{BB2CB8B1-AA48-4361-B6FE-558FE40E8E9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB2CB8B1-AA48-4361-B6FE-558FE40E8E9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB2CB8B1-AA48-4361-B6FE-558FE40E8E9C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E47B721F-EFCA-4302-AAB1-6AE97A62F813}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E47B721F-EFCA-4302-AAB1-6AE97A62F813}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E47B721F-EFCA-4302-AAB1-6AE97A62F813}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E47B721F-EFCA-4302-AAB1-6AE97A62F813}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{C572DA49-1F54-4884-AC07-851CF64264DD} = {91B90448-4C57-49AC-9A49-E4919035E1EF}
 		{3B4BF7C5-73FB-4BF8-B64E-A41BEA003266} = {A960764B-242B-4163-B67B-5B28B8206E50}
 		{DA683491-84B7-4617-9FC1-ECFA183E438A} = {A960764B-242B-4163-B67B-5B28B8206E50}
 		{2F08BA6B-CD20-461B-8FCB-1BCC1831453F} = {A960764B-242B-4163-B67B-5B28B8206E50}
@@ -115,6 +114,7 @@ Global
 		{64DF6A89-5736-4AB3-918A-9DA16398B916} = {91B90448-4C57-49AC-9A49-E4919035E1EF}
 		{6052744D-1AD6-4D25-A2E6-4F81FAE44644} = {A960764B-242B-4163-B67B-5B28B8206E50}
 		{BB2CB8B1-AA48-4361-B6FE-558FE40E8E9C} = {91B90448-4C57-49AC-9A49-E4919035E1EF}
+		{E47B721F-EFCA-4302-AAB1-6AE97A62F813} = {91B90448-4C57-49AC-9A49-E4919035E1EF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {624C7BF3-659F-4D2B-8470-C37A28C6B740}
