+++ test/Candoumbe.DataAccess.EFStore.UnitTests/stryker-config.json
@@ -3,8 +3,7 @@
     "project-info": {
       "name": "github.com/candoumbe/DataAccess",
       "module": "Candoumbe.DataAccess.EFStore"
-    },
-    "concurrency": 4
+    }
   }
 
 }
