+++ test/Candoumbe.DataAccess.RavenDb.UnitTests/stryker-config.json
@@ -3,8 +3,7 @@
     "project-info": {
       "name": "github.com/candoumbe/DataAccess",
       "module": "Candoumbe.DataAccess.RavenDb"
-    },
-    "concurrency": 4
+    }
   }
 
 }
