+++ test/Candoumbe.DataAccess.UnitTests/stryker-config.json
@@ -3,8 +3,7 @@
     "project-info": {
       "name": "github.com/candoumbe/DataAccess",
       "module": "Candoumbe.DataAccess"
-    },
-    "concurrency": 4
+    }
   }
 
 }
